New Node
A node with a global prompt text field and 5 random lines.
Put in a global prompt and up to 5 different categories of random line picking.

The syntax is: beautiful, high quality, apple --- worst quality, cherries, red
So '---' is the marker between positive prompt and negative prompt. It should be used in the same line in the random text fields because each line is one concept which gets randomly picked when rendering.

The output text uses the marker --- in an own line.